requirements.txt
pytest==8.4.1
selenium==4.35.0

webdriver-manager~=4.0.2
requests~=2.32.5
allure-pytest==2.15.0
allure-python-commons~=2.15.0
